Only override the LSTM cell bias initializer if it is not trainable.
Deterministic LSTM cells override the bias initializer to set the forget gate
to a vector of ones such that the cell state is not forgotten at the beginning
of training.  Previously, our stochastic LSTM cell also did this when the
bias_initializer was trainable. This is not always desired, so this change
removes that and only applies the override when the bias_initializer is
deterministic.

PiperOrigin-RevId: 298240699